gitserver: Further move code into subpackages
More code moving into subpackages to reduce the loc in the server.go file and have better separation between the actual gRPC/HTTP server code and just generic git/fs code.
