+++ cmd/gitserver/internal/BUILD.bazel
@@ -14,10 +14,12 @@ go_library(
         "observability.go",
         "patch.go",
         "repo_info.go",
+        "search.go",
         "server.go",
         "server_grpc.go",
         "servermetrics.go",
         "serverutil.go",
+        "statesyncer.go",
     ],
     embedsrcs = ["sg_maintenance.sh"],
     importpath = "github.com/sourcegraph/sourcegraph/cmd/gitserver/internal",
@@ -27,6 +29,7 @@ go_library(
         "//cmd/gitserver/internal/accesslog",
         "//cmd/gitserver/internal/common",
         "//cmd/gitserver/internal/executil",
+        "//cmd/gitserver/internal/git",
         "//cmd/gitserver/internal/gitserverfs",
         "//cmd/gitserver/internal/perforce",
         "//cmd/gitserver/internal/sshagent",
@@ -92,6 +95,7 @@ go_test(
     srcs = [
         "cleanup_test.go",
         "list_gitolite_test.go",
+        "main_test.go",
         "server_test.go",
         "serverutil_test.go",
     ],
@@ -107,6 +111,7 @@ go_test(
     deps = [
         "//cmd/gitserver/internal/common",
         "//cmd/gitserver/internal/executil",
+        "//cmd/gitserver/internal/git",
         "//cmd/gitserver/internal/gitserverfs",
         "//cmd/gitserver/internal/perforce",
         "//cmd/gitserver/internal/vcssyncer",
