+++ cmd/gitserver/internal/server_test.go
@@ -5,10 +5,8 @@ import (
 	"container/list"
 	"context"
 	"encoding/json"
-	"flag"
 	"fmt"
 	"io"
-	"math/rand"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
@@ -28,7 +26,6 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
-	"github.com/sourcegraph/log"
 	"github.com/sourcegraph/log/logtest"
 
 	"github.com/sourcegraph/sourcegraph/cmd/gitserver/internal/common"
@@ -503,150 +500,6 @@ func TestServer_handleP4Exec(t *testing.T) {
 	})
 }
 
-func BenchmarkQuickRevParseHeadQuickSymbolicRefHead_packed_refs(b *testing.B) {
-	tmp := b.TempDir()
-
-	dir := filepath.Join(tmp, ".git")
-	gitDir := common.GitDir(dir)
-	if err := os.Mkdir(dir, 0o700); err != nil {
-		b.Fatal(err)
-	}
-
-	masterRef := "refs/heads/master"
-	// This simulates the most amount of work quickRevParseHead has to do, and
-	// is also the most common in prod. That is where the final rev is in
-	// packed-refs.
-	err := os.WriteFile(filepath.Join(dir, "HEAD"), []byte(fmt.Sprintf("ref: %s\n", masterRef)), 0o600)
-	if err != nil {
-		b.Fatal(err)
-	}
-	// in prod the kubernetes repo has a packed-refs file that is 62446 lines
-	// long. Simulate something like that with everything except master
-	masterRev := "4d5092a09bca95e0153c423d76ef62d4fcd168ec"
-	{
-		f, err := os.Create(filepath.Join(dir, "packed-refs"))
-		if err != nil {
-			b.Fatal(err)
-		}
-		writeRef := func(refBase string, num int) {
-			_, err := fmt.Fprintf(f, "%016x%016x%08x %s-%d\n", rand.Uint64(), rand.Uint64(), rand.Uint32(), refBase, num)
-			if err != nil {
-				b.Fatal(err)
-			}
-		}
-		for i := 0; i < 32; i++ {
-			writeRef("refs/heads/feature-branch", i)
-		}
-		_, err = fmt.Fprintf(f, "%s refs/heads/master\n", masterRev)
-		if err != nil {
-			b.Fatal(err)
-		}
-		for i := 0; i < 10000; i++ {
-			// actual format is refs/pull/${i}/head, but doesn't actually
-			// matter for testing
-			writeRef("refs/pull/head", i)
-			writeRef("refs/pull/merge", i)
-		}
-		err = f.Close()
-		if err != nil {
-			b.Fatal(err)
-		}
-	}
-
-	// Exclude setup
-	b.ResetTimer()
-
-	for n := 0; n < b.N; n++ {
-		rev, err := quickRevParseHead(gitDir)
-		if err != nil {
-			b.Fatal(err)
-		}
-		if rev != masterRev {
-			b.Fatal("unexpected rev: ", rev)
-		}
-		ref, err := quickSymbolicRefHead(gitDir)
-		if err != nil {
-			b.Fatal(err)
-		}
-		if ref != masterRef {
-			b.Fatal("unexpected ref: ", ref)
-		}
-	}
-
-	// Exclude cleanup (defers)
-	b.StopTimer()
-}
-
-func BenchmarkQuickRevParseHeadQuickSymbolicRefHead_unpacked_refs(b *testing.B) {
-	tmp := b.TempDir()
-
-	dir := filepath.Join(tmp, ".git")
-	gitDir := common.GitDir(dir)
-	if err := os.Mkdir(dir, 0o700); err != nil {
-		b.Fatal(err)
-	}
-
-	// This simulates the usual case for a repo that HEAD is often
-	// updated. The master ref will be unpacked.
-	masterRef := "refs/heads/master"
-	masterRev := "4d5092a09bca95e0153c423d76ef62d4fcd168ec"
-	files := map[string]string{
-		"HEAD":              fmt.Sprintf("ref: %s\n", masterRef),
-		"refs/heads/master": masterRev + "\n",
-	}
-	for path, content := range files {
-		path = filepath.Join(dir, path)
-		err := os.MkdirAll(filepath.Dir(path), 0o700)
-		if err != nil {
-			b.Fatal(err)
-		}
-		err = os.WriteFile(path, []byte(content), 0o600)
-		if err != nil {
-			b.Fatal(err)
-		}
-	}
-
-	// Exclude setup
-	b.ResetTimer()
-
-	for n := 0; n < b.N; n++ {
-		rev, err := quickRevParseHead(gitDir)
-		if err != nil {
-			b.Fatal(err)
-		}
-		if rev != masterRev {
-			b.Fatal("unexpected rev: ", rev)
-		}
-		ref, err := quickSymbolicRefHead(gitDir)
-		if err != nil {
-			b.Fatal(err)
-		}
-		if ref != masterRef {
-			b.Fatal("unexpected ref: ", ref)
-		}
-	}
-
-	// Exclude cleanup (defers)
-	b.StopTimer()
-}
-
-func runCmd(t *testing.T, dir string, cmd string, arg ...string) string {
-	t.Helper()
-	c := exec.Command(cmd, arg...)
-	c.Dir = dir
-	c.Env = []string{
-		"GIT_COMMITTER_NAME=a",
-		"GIT_COMMITTER_EMAIL=a@a.com",
-		"GIT_AUTHOR_NAME=a",
-		"GIT_AUTHOR_EMAIL=a@a.com",
-	}
-	b, err := c.CombinedOutput()
-	if err != nil {
-		t.Fatalf("%s %s failed: %s\nOutput: %s", cmd, strings.Join(arg, " "), err, b)
-	}
-	return string(b)
-}
-
 func staticGetRemoteURL(remote string) func(context.Context, api.RepoName) (string, error) {
 	return func(context.Context, api.RepoName) (string, error) {
 		return remote, nil
@@ -784,7 +637,7 @@ func TestCloneRepo(t *testing.T) {
 		}
 		time.Sleep(10 * time.Millisecond)
 	}
-	wantRepoSize := dirSize(repoDir.Path("."))
+	wantRepoSize := gitserverfs.DirSize(repoDir.Path("."))
 	assertRepoState(types.CloneStatusCloned, wantRepoSize, err)
 
 	cmdExecDir = repoDir.Path(".")
@@ -972,7 +825,7 @@ func testHandleRepoDelete(t *testing.T, deletedInDB bool) {
 	req := newRequest("GET", "/repo-update", bytes.NewReader(body))
 	s.handleRepoUpdate(rr, req)
 
-	size := dirSize(gitserverfs.RepoDirFromName(s.ReposDir, repoName).Path("."))
+	size := gitserverfs.DirSize(gitserverfs.RepoDirFromName(s.ReposDir, repoName).Path("."))
 	want := &types.GitserverRepo{
 		RepoID:        dbRepo.ID,
 		ShardID:       "",
@@ -1014,7 +867,7 @@ func testHandleRepoDelete(t *testing.T, deletedInDB bool) {
 	req = newRequest("GET", "/delete", bytes.NewReader(body))
 	s.handleRepoDelete(rr, req)
 
-	size = dirSize(gitserverfs.RepoDirFromName(s.ReposDir, repoName).Path("."))
+	size = gitserverfs.DirSize(gitserverfs.RepoDirFromName(s.ReposDir, repoName).Path("."))
 	if size != 0 {
 		t.Fatalf("Size should be 0, got %d", size)
 	}
@@ -1089,7 +942,7 @@ func TestHandleRepoUpdate(t *testing.T) {
 	req := newRequest("GET", "/repo-update", bytes.NewReader(body))
 	s.handleRepoUpdate(rr, req)
 
-	size := dirSize(gitserverfs.RepoDirFromName(s.ReposDir, repoName).Path("."))
+	size := gitserverfs.DirSize(gitserverfs.RepoDirFromName(s.ReposDir, repoName).Path("."))
 	want := &types.GitserverRepo{
 		RepoID:        dbRepo.ID,
 		ShardID:       "",
@@ -1119,7 +972,7 @@ func TestHandleRepoUpdate(t *testing.T) {
 	req = newRequest("GET", "/repo-update", bytes.NewReader(body))
 	s.handleRepoUpdate(rr, req)
 
-	size = dirSize(gitserverfs.RepoDirFromName(s.ReposDir, repoName).Path("."))
+	size = gitserverfs.DirSize(gitserverfs.RepoDirFromName(s.ReposDir, repoName).Path("."))
 	want = &types.GitserverRepo{
 		RepoID:        dbRepo.ID,
 		ShardID:       "",
@@ -1175,7 +1028,7 @@ func TestHandleRepoUpdate(t *testing.T) {
 		RepoID:        dbRepo.ID,
 		ShardID:       "",
 		CloneStatus:   types.CloneStatusCloned,
-		RepoSizeBytes: dirSize(gitserverfs.RepoDirFromName(s.ReposDir, repoName).Path(".")), // we compute the new size
+		RepoSizeBytes: gitserverfs.DirSize(gitserverfs.RepoDirFromName(s.ReposDir, repoName).Path(".")), // we compute the new size
 	}
 	fromDB, err = db.GitserverRepos().GetByID(ctx, dbRepo.ID)
 	if err != nil {
@@ -1185,51 +1038,6 @@ func TestHandleRepoUpdate(t *testing.T) {
 	// We expect an update
 	if diff := cmp.Diff(want, fromDB, ignoreVolatileGitserverRepoFields); diff != "" {
 		t.Fatal(diff)
-	}
-}
-
-func TestRemoveBadRefs(t *testing.T) {
-	dir := t.TempDir()
-	gitDir := common.GitDir(filepath.Join(dir, ".git"))
-
-	cmd := func(name string, arg ...string) string {
-		t.Helper()
-		return runCmd(t, dir, name, arg...)
-	}
-	wantCommit := makeSingleCommitRepo(cmd)
-
-	for _, name := range []string{"HEAD", "head", "Head", "HeAd"} {
-		// Tag
-		cmd("git", "tag", name)
-
-		if dontWant := cmd("git", "rev-parse", "HEAD"); dontWant == wantCommit {
-			t.Logf("WARNING: git tag %s failed to produce ambiguous output: %s", name, dontWant)
-		}
-
-		if err := removeBadRefs(context.Background(), gitDir); err != nil {
-			t.Fatal(err)
-		}
-
-		if got := cmd("git", "rev-parse", "HEAD"); got != wantCommit {
-			t.Fatalf("git tag %s failed to be removed: %s", name, got)
-		}
-
-		// Ref
-		if err := os.WriteFile(filepath.Join(dir, ".git", "refs", "heads", name), []byte(wantCommit), 0o600); err != nil {
-			t.Fatal(err)
-		}
-
-		if dontWant := cmd("git", "rev-parse", "HEAD"); dontWant == wantCommit {
-			t.Logf("WARNING: git ref %s failed to produce ambiguous output: %s", name, dontWant)
-		}
-
-		if err := removeBadRefs(context.Background(), gitDir); err != nil {
-			t.Fatal(err)
-		}
-
-		if got := cmd("git", "rev-parse", "HEAD"); got != wantCommit {
-			t.Fatalf("git ref %s failed to be removed: %s", name, got)
-		}
 	}
 }
 
@@ -1822,12 +1630,32 @@ func mustEncodeJSONResponse(value any) string {
 	return strings.TrimSpace(string(encoded))
 }
 
-func TestMain(m *testing.M) {
-	flag.Parse()
-	if !testing.Verbose() {
-		logtest.InitWithLevel(m, log.LevelNone)
-	} else {
-		logtest.Init(m)
+func TestStdErrIndicatesCorruption(t *testing.T) {
+	bad := []string{
+		"error: packfile .git/objects/pack/pack-a.pack does not match index",
+		"error: Could not read d24d09b8bc5d1ea2c3aa24455f4578db6aa3afda\n",
+		`error: short SHA1 1325 is ambiguous
+error: Could not read d24d09b8bc5d1ea2c3aa24455f4578db6aa3afda`,
+		`unrelated
+error: Could not read d24d09b8bc5d1ea2c3aa24455f4578db6aa3afda`,
+		"\n\nerror: Could not read d24d09b8bc5d1ea2c3aa24455f4578db6aa3afda",
+		"fatal: commit-graph requires overflow generation data but has none\n",
+		"\rResolving deltas: 100% (21750/21750), completed with 565 local objects.\nfatal: commit-graph requires overflow generation data but has none\nerror: https://github.com/sgtest/megarepo did not send all necessary objects\n\n\": exit status 1",
 	}
-	os.Exit(m.Run())
+	good := []string{
+		"",
+		"error: short SHA1 1325 is ambiguous",
+		"error: object 156639577dd2ea91cdd53b25352648387d985743 is a blob, not a commit",
+		"error: object 45043b3ff0440f4d7937f8c68f8fb2881759edef is a tree, not a commit",
+	}
+	for _, stderr := range bad {
+		if !stdErrIndicatesCorruption(stderr) {
+			t.Errorf("should contain corrupt line:\n%s", stderr)
+		}
+	}
+	for _, stderr := range good {
+		if stdErrIndicatesCorruption(stderr) {
+			t.Errorf("should not contain corrupt line:\n%s", stderr)
+		}
+	}
 }
